+++ apps/voxelize.cpp
@@ -33,11 +33,9 @@
 
 #include <fivox/fivox.h>
 #include <fivox/beerLambertProjectionImageFilter.h>
+#include <fivox/scaleFilter.h>
 
-#include <itkCastImageFilter.h>
 #include <itkImageFileWriter.h>
-#include <itkRescaleIntensityImageFilter.h>
-#include <itkShiftScaleImageFilter.h>
 
 #include <lunchbox/file.h>
 #include <lunchbox/log.h>
@@ -55,46 +53,26 @@ typedef itk::Image< FloatPixelType, 2 > FloatImageType;
 
 template< typename T > class VolumeWriter
 {
-    typedef itk::RescaleIntensityImageFilter< fivox::FloatVolume,
-                                              itk::Image< T, 3 >> RescaleFilter;
-    typedef itk::ShiftScaleImageFilter< fivox::FloatVolume,
-                                        fivox::FloatVolume > ScaleFilter;
-    typedef itk::CastImageFilter< fivox::FloatVolume,
-                                  itk::Image< T, 3 >> CastFilter;
     typedef itk::ImageFileWriter< itk::Image< T, 3 >> Writer;
+    typedef fivox::ScaleFilter< T > ScaleFilter;
 
 public:
-    VolumeWriter( VolumePtr input, const bool rescale )
-        : _writer( Writer::New( ))
+    VolumeWriter( VolumePtr input, const vmml::Vector2f& dataRange )
+        : _scaler( input, dataRange )
+        , _writer( Writer::New( ))
     {
-        if( rescale )
-        {
-            _rescale = RescaleFilter::New();
-            _rescale->SetInput( input );
-            _writer->SetInput( _rescale->GetOutput( ));
-        }
-        else
-        {
-            _scale = ScaleFilter::New();
-            _cast = CastFilter::New();
-
-            _scale->SetInput( input );
-            _scale->SetScale( std::numeric_limits< T >::max( ));
-            _cast->SetInput( _scale->GetOutput( ));
-            _writer->SetInput( _cast->GetOutput( ));
-        }
+        _writer->SetInput( _scaler->GetOutput( ));
     }
 
     typename Writer::Pointer operator->() { return _writer; }
 
 private:
-    typename RescaleFilter::Pointer _rescale;
-    typename ScaleFilter::Pointer _scale;
-    typename CastFilter::Pointer _cast;
+    ScaleFilter _scaler;
     typename Writer::Pointer _writer;
 };
 
-template<> VolumeWriter< float >::VolumeWriter( VolumePtr input, bool )
+template<> VolumeWriter< float >::VolumeWriter( VolumePtr input,
+                                                const vmml::Vector2f& )
     : _writer( Writer::New( ))
 {
     _writer->SetInput( input );
@@ -102,11 +80,11 @@ template<> VolumeWriter< float >::VolumeWriter( VolumePtr input, bool )
 
 template< typename T >
 void _sample( ImageSourcePtr source, const vmml::Vector2ui& frameRange,
-              const double sigmaVSDProjection, const float volumeResolution,
-              const std::string& outputFile, const bool rescale )
+              const double sigmaVSDProjection, const fivox::URIHandler& params,
+              const std::string& outputFile )
 {
     VolumePtr input = source->GetOutput();
-    VolumeWriter< T > writer( input, rescale );
+    VolumeWriter< T > writer( input, params.getInputRange( ));
 
     const size_t numDigits = std::to_string( frameRange.y( )).length();
     for( uint32_t i = frameRange.x(); i < frameRange.y(); ++i )
@@ -139,7 +117,7 @@ void _sample( ImageSourcePtr source, const vmml::Vector2ui& frameRange,
         FilterType::Pointer projection = FilterType::New();
         projection->SetInput( input );
         projection->SetProjectionDimension( 1 ); // projection along Y-axis
-        projection->SetPixelSize( 1.0 / volumeResolution );
+        projection->SetPixelSize( 1.0 / params.getResolution( ));
         projection->SetSigma( sigmaVSDProjection );
 
         // Write output image
@@ -204,18 +182,19 @@ int main( int argc, char* argv[] )
           "\n"
           "Parameters for all types :\n"
           "- BlueConfig: BlueConfig file path\n"
-          "              (default: 'configs/BlueConfigVSD' for VSD reports,\n"
-          "               BBPTestData otherwise)\n"
+          "              (default: BBPTestData)\n"
           "- target: name of the BlueConfig target (default: CircuitTarget)\n"
-          "- magnitude: value multiplied to each sampled voxel value\n"
-          "             (defaults: 0.1 for Compartments and Somas,\n"
-          "                        1.5 / duration for Spikes\n"
-          "                        1.0 for Synapses and VSD)\n"
+          "- inputMin/inputMax: minimum and maximum input values to be considered for rescaling\n"
+          "                     (defaults: [0.0, 2.0] for Spikes and Synapses\n"
+          "                                [-190.0, 0.0] for Compartments with TestData, [-80.0, 0.0] otherwise\n"
+          "                                [-15.0, 0.0] for Somas with TestData, [-80.0, 0.0] otherwise\n"
+          "                                [-0.0000147, 0.00225] for LFP with TestData, [-10.0, 10.0] otherwise\n"
+          "                                [-100000.0, 300.0] for VSD)\n"
           "- functor: type of functor to sample the data into the voxels\n"
           "             (defaults: \"density\" for Synapses,\n"
           "                        \"frequency\" for Spikes,\n"
           "                        \"field\" for Compartments, Somas and VSD)\n"
-          "- resolution: number of voxels per micrometer (default: 1.0)\n"
+          "- resolution: number of voxels per micrometer (default: 10.0)\n"
           "- maxBlockSize: maximum memory usage allowed for one block in bytes\n"
           "                (default: 64MB)\n"
           "- maxError: maximum error allowed (default: 0.001). If the event's\n"
@@ -319,13 +298,16 @@ int main( int argc, char* argv[] )
     ::fivox::URIHandler params( uri );
     ImageSourcePtr source = params.newImageSource< float >();
     ::fivox::EventSourcePtr loader = source->getFunctor()->getSource();
+
     const fivox::AABBf& bbox = loader->getBoundingBox();
-    const fivox::Vector3f& position = bbox.getMin();
-    const float extent = bbox.getSize().find_max();
+    const fivox::Vector3f& extent( bbox.getSize() +
+                                   loader->getCutOffDistance() * 2.f );
+    const float maxExtent = extent.find_max();
 
     fivox::FloatVolume::SizeType vSize;
-    vSize.Fill( size );
-    vSize[2] = end - begin + 1;
+    vSize[0] = size * extent[0] / maxExtent;
+    vSize[1] = size * extent[1] / maxExtent;
+    vSize[2] = ( end - begin + 1 ) * extent[2] / maxExtent ;
 
     fivox::FloatVolume::IndexType vIndex;
     vIndex.Fill( 0 );
@@ -339,9 +321,10 @@ int main( int argc, char* argv[] )
     output->SetRegions( region );
 
     typename fivox::FloatVolume::SpacingType spacing;
-    spacing.Fill( extent / float( size ));
+    spacing.Fill( maxExtent / float( size ));
     output->SetSpacing( spacing );
 
+    const fivox::Vector3f& position( bbox.getCenter() - extent * 0.5f );
     typename fivox::FloatVolume::PointType origin;
     origin[0] = position[0];
     origin[1] = position[1];
@@ -372,33 +355,29 @@ int main( int argc, char* argv[] )
             params.getType() == fivox::TYPE_VSD && vm.count( "projection" ) ?
                 vm["projection"].as< double >() : -1.0;
 
-    // Only rescale/normalize float input for single, full volumes
-    const bool rescale = frameRange[1] - frameRange[0] == 1 &&
-                         decompose[1] == 1;
-
     const std::string& datatype( vm["datatype"].as< std::string >( ));
     if( datatype == "char" )
     {
         LBINFO << "Sampling volume as char (uint8_t) data" << std::endl;
         _sample< uint8_t >( source, frameRange, sigmaVSDProjection,
-                            params.getResolution(), outputFile, rescale );
+                            params, outputFile );
     }
     else if( datatype == "short" )
     {
         LBINFO << "Sampling volume as short (uint16_t) data" << std::endl;
         _sample< uint16_t >( source, frameRange, sigmaVSDProjection,
-                             params.getResolution(), outputFile, rescale );
+                             params, outputFile );
     }
     else if( datatype == "int" )
     {
         LBINFO << "Sampling volume as int (uint32_t) data" << std::endl;
         _sample< uint32_t >( source, frameRange, sigmaVSDProjection,
-                             params.getResolution(), outputFile, rescale );
+                             params, outputFile );
     }
     else
     {
         LBINFO << "Sampling volume as floating point data" << std::endl;
         _sample< float >( source, frameRange, sigmaVSDProjection,
-                          params.getResolution(), outputFile, false );
+                          params, outputFile );
     }
 }