+++ fivox/compartmentLoader.cpp
@@ -47,13 +47,21 @@ public:
         , _config( params.getConfig( ))
         , _target( _config.parseTarget( params.getTarget( )))
         , _report( _config.getReportSource( params.getReport( )),
-                   brion::MODE_READ, _target)
+                   brion::MODE_READ, _target )
     {
         brain::Circuit circuit( _config );
         const auto morphologies = circuit.loadMorphologies(
             _target, brain::Circuit::COORDINATES_GLOBAL );
 
         helpers::addCompartmentEvents( morphologies, _report, output );
+
+        const float max = -60.f;
+        const float distance =
+                std::sqrt( std::abs( max ) / params.getMaxError( ));
+        LBINFO << "Computed cutoff distance: " << distance
+               << " with maximum event's value: " << max << std::endl;
+
+        output.setCutOffDistance( distance );
     }
 
     ssize_t load( const float time )
@@ -62,7 +70,7 @@ public:
         if( !values )
             return -1;
 
-        for( size_t i = 0; i != values->size( ); ++i )
+        for( size_t i = 0; i != values->size(); ++i )
             _output[i].value = ( *values )[i];
 
         return values->size();