+++ fivox/densityFunctor.h
@@ -36,7 +36,9 @@ class DensityFunctor : public EventFunctor< TImage >
     typedef typename Super::TSpacing TSpacing;
 
 public:
-    DensityFunctor() {}
+    DensityFunctor( const fivox::Vector2f& inputRange )
+        : Super( inputRange )
+    {}
     virtual ~DensityFunctor() {}
 
     TPixel operator()( const TPoint& point, const TSpacing& spacing )
@@ -65,6 +67,7 @@ DensityFunctor< TImage >::operator()( const TPoint& itkPoint,
     float sum = 0.f;
     for( const Event& event : events )
         sum += event.value;
+
     sum /= std::abs( spacing_2.product() * 8.f );
     return Super::_scale( sum );
 }