+++ fivox/fieldFunctor.h
@@ -35,43 +35,13 @@ template< typename TImage > class FieldFunctor : public EventFunctor< TImage >
     typedef typename Super::TSpacing TSpacing;
 
 public:
-    FieldFunctor( const float magnitude, const float maxError )
-        : _cutOffDistance( 50. )
-        , _magnitude( magnitude )
-        , _maxError( maxError )
+    FieldFunctor( const fivox::Vector2f& inputRange )
+        : Super( inputRange )
     {}
     virtual ~FieldFunctor() {}
 
-    void beforeGenerate() override
-    {
-        Super::beforeGenerate();
-        float max = brion::MINIMUM_VOLTAGE;
-        for( const Event& event : Super::_source->getEvents( ))
-        {
-            if( event.value == VALUE_UNSET )
-                continue;
-
-            const float eventValue =
-                    ( event.value - brion::MINIMUM_VOLTAGE ) * _magnitude;
-            max = std::max( max, eventValue );
-        }
-
-        const float distance = std::sqrt( max / _maxError );
-        if( _cutOffDistance != distance )
-        {
-            LBINFO << "Computed cutoff distance: " << distance
-                   << " with maximum event's value: " << max << std::endl;
-            _cutOffDistance = distance;
-        }
-    }
-
     TPixel operator()( const TPoint& point, const TSpacing& spacing )
         const override;
-
-private:
-    float _cutOffDistance;
-    const float _magnitude;
-    const float _maxError;
 };
 
 template< class TImage > inline typename FieldFunctor< TImage >::TPixel
@@ -85,16 +55,15 @@ FieldFunctor< TImage >::operator()( const TPoint& point, const TSpacing& ) const
     for( size_t i = 0; i < components; ++i )
         base[i] = point[i];
 
-    const AABBf region( base - Vector3f( _cutOffDistance ),
-                        base + Vector3f( _cutOffDistance ));
+    const float cutOffDistance = Super::_source->getCutOffDistance();
+    const AABBf region( base - Vector3f( cutOffDistance ),
+                        base + Vector3f( cutOffDistance ));
     const Events& events = Super::_source->findEvents( region );
 
-    float sum = 0.f;
+    const float squaredCutoff = cutOffDistance * cutOffDistance;
+    float sum = 0;
     for( const Event& event : events )
     {
-        const float eventValue =
-                ( event.value - brion::MINIMUM_VOLTAGE ) * _magnitude;
-
         // OPT: do 'manual' operator- and squared_length(), vtune says it's
         // faster than using vmml vector functions
         const Vector3f distance( base.array[0] - event.position.array[0],
@@ -104,12 +73,16 @@ FieldFunctor< TImage >::operator()( const TPoint& point, const TSpacing& ) const
                                distance.array[1] * distance.array[1] +
                                distance.array[2] * distance.array[2] );
 
-        if( distance2 > 1. )
-            sum += eventValue / distance2;
-        else
-            sum += eventValue;
-    }
+        if( distance2 > squaredCutoff )
+            continue;
 
+        // If center of the voxel within the event radius, use the
+        // voltage at the surface of the compartment (at 'radius' distance)
+        const float contribution =
+                distance2 < event.radius * event.radius ? 1.f / event.radius
+                                                        : 1.f / distance2;
+        sum += contribution * event.value;
+    }
     return Super::_scale( sum );
 }