+++ tests/sources.cpp
@@ -35,6 +35,7 @@
 #include "test.h"
 #include <fivox/compartmentLoader.h>
 #include <fivox/eventFunctor.h>
+#include <fivox/helpers.h>
 #include <fivox/imageSource.h>
 #ifdef FIVOX_USE_LFP
 #  include <fivox/lfp/lfpFunctor.h>
@@ -86,7 +87,7 @@ inline float _testKernel(
 
     // set up size and origin for loaded data
     fivox::EventSourcePtr source = filter->getFunctor()->getSource();
-    source->load( 5.f );
+    source->load( 0.f );
     const fivox::AABBf& bbox = source->getBoundingBox();
     const fivox::Vector3f& position = bbox.getMin();
     const float extent = bbox.getSize().find_max();
@@ -211,7 +212,15 @@ BOOST_AUTO_TEST_CASE( fivoxVoltages_source )
 {
     // Compartment report 'voltages' (binary) contains timestamps
     // between 0 and 100 with a Dt=0.1 => data range is 0.0 to 10.0 ms
-    testSource( "fivox://", 0.54296875f, 0.0022559839199516318f,
+    testSource( "fivox://", 254.529296875f, -0.26330676218117333f,
+                vmml::Vector2ui( 0, 100 ));
+}
+
+BOOST_AUTO_TEST_CASE( fivoxSomas_source )
+{
+    // Soma report 'somas' (binary) contains timestamps
+    // between 0 and 100 with a Dt=0.1 => data range is 0.0 to 10.0 ms
+    testSource( "fivoxSomas://", 254.927734375f, -0.0021073255409191916f,
                 vmml::Vector2ui( 0, 100 ));
 }
 
@@ -220,33 +229,22 @@ BOOST_AUTO_TEST_CASE( fivoxLFP_source )
 {
     // Compartment currents report 'currents' (binary) contains timestamps
     // between 0 and 100 with a Dt=0.1 => data range is 0.0 to 10.0 ms
-    testSource( "fivoxlfp://", 0.f, 6.7447619018805392e-05,
+    testSource( "fivoxlfp://", 2.01953125f, 8.9861975556847204e-06f,
                 vmml::Vector2ui( 0, 100 ));
 }
 #endif
 
-BOOST_AUTO_TEST_CASE( fivoxSomas_source )
-{
-    // Soma report 'somas' (binary) contains timestamps
-    // between 0 and 100 with a Dt=0.1 => data range is 0.0 to 10.0 ms
-    testSource( "fivoxSomas://", 0.005859375f, 4.3455195964270388e-05f,
-                vmml::Vector2ui( 0, 100 ));
-}
-
 BOOST_AUTO_TEST_CASE( fivoxSpikes_source )
 {
     // Spikes report timestamps range between 0.725 and 9.975 ms
-    // Better, but not always available:
-    // "fivoxSpikes:///gpfs/bbp.cscs.ch/home/nachbaur/BlueConfig_3m",
     testSource( "fivoxSpikes://?duration=1,dt=1",
-                2.490234375f, 0.0146484375f, vmml::Vector2ui( 0, 9 ));
-
+                0.49609375f, 0.00390625f, vmml::Vector2ui( 0, 9 ));
 }
 
 BOOST_AUTO_TEST_CASE( fivoxSynapses_source )
 {
     // Synapse reports don't have time support and return a 1-frame interval
-    testSource( "fivoxSynapses://", 0.03515625f, 0.00017834029219887526f,
+    testSource( "fivoxSynapses://", 0.f, 1.7834029313844313e-05f,
                 vmml::Vector2ui( 0, 1 ));
 }