+++ doc/Changelog.md
@@ -3,6 +3,10 @@ Changelog {#changelog}
 
 # git master {#master}
 
+* [#10](https://github.com/BlueBrain/Fivox/pull/10)
+  voxelize: Limit volume to area covered by events.
+* [#10](https://github.com/BlueBrain/Fivox/pull/10)
+  Rescale to the output type range for integer volumes.
 * [#9](https://github.com/BlueBrain/Fivox/pull/9)
   Add testLoader for validation purposes.
 * [#4](https://github.com/BlueBrain/Fivox/pull/4)
@@ -12,7 +16,7 @@ Changelog {#changelog}
   Add radius attribute in Event, used in the CompartmentLoader.
   Adapt tests and default values to latest TestData (370ebee).
 * [#7](https://github.com/BlueBrain/Fivox/pull/7):
-  Add data decomposition in voxelize tool
+  Add data decomposition in voxelize tool.
 * [20437](https://bbpcode.epfl.ch/code/#/c/20437/)
   Expose the VSD projection attenuation coefficient (absorption + scattering)
   in the voxelize command line tool.